[refactor] #43 post sort 이름 변경
